Initial implementation of SmallPID.
This is 100% based on FastPID but PID parameters are calculated (from
float to int) at compile time and stored as constexpr. This of course
rules out online parameter tuning, but if the parameters are finally
known the code size might be reduced greatly.

The reason is that while PID parameters aren't known it's a good idea
to use FastPID implementation as the PID params could be set at run
time. However once the process is done a switch to SmallPID will
reduce the code for production.

The configuration of SmallPID might look like this:
```c++
using cfgT = PID::Cfg<0.2f, 0.4f, 0.5f, 1.f>;
using pidT = PID::SmallPID<cfgT>;
pidT myPid(bits, sign);
```

**Warning** keep in mind that this code compiles only if c++20 support
  is available.

Further possible extensions, not yet done, would be to make the output
range as well as bits and sign constexpr as well. However I doubt this
would reduce the code significantly.